e2e(template): move templates related files in e2e/templates folder
