--- e2e/.gitlab.env
@@ -1,1 +1,0 @@
-../templates/simple/environments/.gitlab.env
