--- e2e/helpers.ts
@@ -1,13 +1,0 @@
-import { config } from "dotenv";
-import { resolve } from "path";
-
-export const gitlabEnv = config({ path: resolve(__dirname, "./.gitlab.env") })
-  .parsed;
-export const BIN_FOLDER = resolve(__dirname, "../node_modules/.bin");
-export const KOSKO_BIN = resolve(BIN_FOLDER, "kosko");
-
-export const TIMEOUT = 25_000; // = 1000 * 25s
-
-export const template = (testFilename: string): string => {
-  return resolve(__dirname, "../templates/", testFilename);
-};
