+++ e2e/templates/nginx/kosko generate/__tests__/--env prod.ts
@@ -1,7 +1,7 @@
 //
 
 import { config } from "dotenv";
-import { KOSKO_BIN, template, TIMEOUT } from "e2e/helpers";
+import { KOSKO_BIN, template, TIMEOUT } from "e2e/templates/helpers";
 // eslint-disable-next-line import/default
 import execa from "execa";
 import { basename, resolve } from "path";
