+++ e2e/templates/pgweb/kosko generate/__tests__/--env dev.ts
@@ -1,7 +1,7 @@
 //
 
 import { config } from "dotenv";
-import { KOSKO_BIN, template, TIMEOUT } from "e2e/helpers";
+import { KOSKO_BIN, template, TIMEOUT } from "e2e/templates/helpers";
 // eslint-disable-next-line import/default
 import execa from "execa";
 import { basename, resolve } from "path";
